+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- JWT kann als OIDC Access Token konfiguriert werden (hitobito_sac_cas#1110)
 - Unterstützung vom login=prompt Parameter im OIDC Authorization flow (hitobito_sac_cas#1075)
 
 ## Version 2.2